# ai-recruiting-platform/packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI Recruiting Platform API
  version: 1.0.0
  description: Production-grade AI recruiting platform with automated sourcing, scoring, and outreach
servers:
  - url: https://api.recruiting-platform.com/v1
    description: Production server
  - url: https://staging-api.recruiting-platform.com/v1
    description: Staging server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
  /jobs:
    post:
      summary: Create a new job from JD
      operationId: jobs.create
      tags:
        - Jobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /candidates:
    post:
      summary: Ingest candidate from resume or profile
      operationId: candidates.ingest
      tags:
        - Candidates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateIngestRequest'
      responses:
        '201':
          description: Candidate ingested successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
...
  /webhooks/outreach:
    post:
      summary: Webhook for outreach events
      operationId: webhooks.outreach
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
      responses:
        '200':
          description: Webhook processed successfully
  /webhooks/calendar:
    post:
      summary: Webhook for calendar events
      operationId: webhooks.calendar
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookEvent'
      responses:
        '200':
          description: Webhook processed successfully
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    JobCreateRequest:
      type: object
      required:
        - jdText
      properties:
        jdText:
          type: string
          description: Job description text
        fileUrl:
          type: string
          description: URL to JD document (PDF/DOC)
        title:
          type: string
          description: Job title (auto-extracted if not provided)
        location:
          type: string
          description: Job location
        compensation:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
            currency:
              type: string
    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: string
        requirements:
          type: array
          items:
            type: string
        mustHaves:
          type: array
          items:
            type: string
        niceToHaves:
          type: array
          items:
            type: string
        minYears:
          type: number
        compensation:
          type: object
          properties:
            min:
              type: number
            max:
              type: number
            currency:
              type: string
        status:
          type: string
          enum: [draft, published, closed]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CandidateIngestRequest:
      type: object
      required:
        - jobId
      properties:
        jobId:
          type: string
        fileUrl:
          type: string
        emailParse:
          type: object
          properties:
            from:
              type: string
            subject:
              type: string
            body:
              type: string
        profile:
          type: object
          properties:
            linkedinUrl:
              type: string
            githubUrl:
              type: string
            stackoverflowUrl:
              type: string
    Candidate:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        currentRole:
          type: string
        currentCompany:
          type: string
        yearsExperience:
          type: number
        skills:
          type: array
          items:
            type: string
        education:
          type: array
          items:
            type: object
        experience:
          type: array
          items:
            type: object
        resumeUrl:
          type: string
        linkedinUrl:
          type: string
        githubUrl:
          type: string
        stackoverflowUrl:
          type: string
        source:
          type: string
        stage:
          type: string
          enum: [new, screening, shortlist, interview, offer, hired, rejected]
        fitScore:
          type: number
        coverage:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WebhookEvent:
      type: object
      required:
        - type
        - timestamp
        - data
      properties:
        type:
          type: string
          enum: [outreach.opened, outreach.replied, outreach.bounced, calendar.created, calendar.updated, ats.sync]
        timestamp:
          type: string
          format: date-time
        data:
          type: object
        signature:
          type: string
          description: HMAC signature for verification
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object